+++ src/darker/command_line.py
@@ -34,6 +34,12 @@ def parse_command_line(argv: List[str]) -> Namespace:
         "-i", "--isort", action="store_true", help="".join(isort_help),
     )
     parser.add_argument(
+        "-c",
+        "--config",
+        metavar="PATH",
+        help="Ask `black` to read configuration from PATH.",
+    )
+    parser.add_argument(
         "-v",
         "--verbose",
         dest="log_level",
@@ -51,5 +57,19 @@ def parse_command_line(argv: List[str]) -> Namespace:
     )
     parser.add_argument(
         "--version", action="store_true", help="Show the version of `darker`"
+    )
+    parser.add_argument(
+        "-S",
+        "--skip-string-normalization",
+        action="store_true",
+        dest="skip_string_normalization",
+        help="Don't normalize string quotes or prefixes",
+    )
+    parser.add_argument(
+        "-l",
+        "--line-length",
+        type=int,
+        dest="line_length",
+        help="How many characters per line to allow [default: 88]",
     )
     return parser.parse_args(argv)