+++ src/darker/tests/conftest.py
@@ -1,8 +1,11 @@
 import sys
 import types
+from subprocess import check_call
+from typing import Dict
 from unittest.mock import patch
 
 import pytest
+from py.path import local as LocalPath
 
 
 @pytest.fixture
@@ -15,3 +18,28 @@ def without_isort():
 def with_isort():
     with patch.dict(sys.modules, {"isort": types.ModuleType("isort")}):
         yield
+
+
+class GitRepoFixture:
+    def __init__(self, root: LocalPath):
+        self.root = root
+
+    def add(
+        self, paths_and_contents: Dict[str, str], commit: str = None
+    ) -> Dict[str, LocalPath]:
+        absolute_paths = {
+            relative_path: self.root / relative_path
+            for relative_path in paths_and_contents
+        }
+        for relative_path, content in paths_and_contents.items():
+            absolute_paths[relative_path].write(content)
+            check_call(["git", "add", relative_path], cwd=self.root)
+        if commit:
+            check_call(["git", "commit", "-m", commit], cwd=self.root)
+        return absolute_paths
+
+
+@pytest.fixture
+def git_repo(tmpdir):
+    check_call(["git", "init"], cwd=tmpdir)
+    return GitRepoFixture(tmpdir)