+++ src/darker/tests/test_black_diff.py
@@ -1,8 +1,10 @@
+from pathlib import Path
 from textwrap import dedent
 
+import pytest
 from black import FileMode, format_str
 
-from darker.black_diff import diff_and_get_opcodes, opcodes_to_chunks
+from darker.black_diff import diff_and_get_opcodes, opcodes_to_chunks, read_black_config
 
 FUNCTIONS2_PY = dedent(
     '''\
@@ -147,3 +149,34 @@ def test_mixed():
             ['    print("Inner defs should breathe a little.")'],
         ),
     ]
+
+
+@pytest.mark.parametrize(
+    "config_path, config_lines, expect",
+    [
+        (None, ['line-length = 79'], {'line_length': 79}),
+        ("custom.toml", ['line-length = 99'], {'line_length': 99}),
+        (
+            "custom.toml",
+            ['skip-string-normalization = true'],
+            {'skip_string_normalization': True},
+        ),
+        (
+            "custom.toml",
+            ['skip-string-normalization = false'],
+            {'skip_string_normalization': False},
+        ),
+        ("custom.toml", ["target-version = ['py37']"], {}),
+        ("custom.toml", ["include = '\\.pyi$'"], {}),
+        ("custom.toml", ["exclude = '\\.pyx$'"], {}),
+    ],
+)
+def test_black_config(tmpdir, config_path, config_lines, expect):
+    tmpdir = Path(tmpdir)
+    src = tmpdir / "src.py"
+    toml = tmpdir / (config_path or "pyproject.toml")
+
+    toml.write_text("[tool.black]\n{}\n".format('\n'.join(config_lines)))
+
+    config = read_black_config(src, config_path and str(toml))
+    assert config == expect