+++ CHANGES.rst
@@ -1,3 +1,9 @@
+0.3.0.dev / yyyy-mm-dd
+----------------------
+
+- Feature: Add support for black config
+- Feature: Add support for ``-l``/``--line-length`` and ``-S``/``--skip-string-normalization``
+
 0.2.0 / 2020-03-11
 ------------------